# pkg/kitchen_api/kitchen_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "kitchen_api.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/orders": {
      "get": {
        "operationId": "Kitchen_GetCookOrders",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetCookOrdersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Kitchen"
        ]
      }
    },
    "/v1/orders/{order_id}/cooking/end": {
      "get": {
        "operationId": "Kitchen_CookingEnd",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CookingEndResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Kitchen"
        ]
      }
    },
    "/v1/orders/{order_id}/cooking/start": {
      "get": {
        "operationId": "Kitchen_CookingStart",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CookingStartResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Kitchen"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1CookingEndResponse": {
      "type": "object"
    },
    "v1CookingStartResponse": {
      "type": "object"
    },
    "v1GetCookOrdersResponse": {
      "type": "object",
      "properties": {
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Order"
          }
        },
        "total": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1KitchenOrderStatus": {
      "type": "string",
      "enum": [
        "KitchenOrderStatusInvalid",
        "kitchenOrderStatusNew",
        "kitchenOrderStatusCooking",
        "kitchenOrderStatusCooked"
      ],
      "default": "KitchenOrderStatusInvalid"
    },
    "v1Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "$ref": "#/definitions/v1KitchenOrderStatus"
        },
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1OrderItem"
          }
        }
      }
    },
    "v1OrderItem": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        }
      }
    }
  }
}